Split init.vim and make it a loader.
The init.vim is only used on non nix systems.  It is used to load
the init_not_nixos.vim which is also only used on non nix based systems.
Otherwise, everything is as it was before.

Note that the plugins/ and load_plugins.vim is only used on non nix
systems, but the other files are used on nix systems, to enable
mostly common systems to be used.