Make nvim config cross-compat with Nix and Non-nix
The nix config for neovim starts in .config/nvim/vim.nix.
This bypasses the init.vim as it's written by the
`programs.neovim` config item.  The plugins.vim files
are modified so that vim-plug is *not* used if running
under nix, as the nix version uses it's own package-manager
to add all the submodules.

So currently, this is compatible between both Nix and
non-Nix (Home Manager) configurations.