Add declarative Nix configuration for LSP servers
This uses the common elements of the interactive configuration
that nvim-lsp-installer uses, and then implements (in servers.lua)
a declarative version for the servers that exist in the
vim.nix (also added).

This means that when vim.nix is included, that all the servers
are installed and configured.  No Vim commands need to be
run to install the LSP servers as it's all done via
Nix.

The config is still compatible with nvim-lsp-installer when this
repo is used as a vcsh module.