Add rustacean (for rust) config and inlay hints
Quite a bit of munging stuff around.  Rustacean is configured
differently to LSP plugins in that it kind of runs it separtely; hence
we have to configure it outside of lsp_config, but we still want to use
the handlers.

Also add in the inlay_hints; essentially, we add <leader>gh to toggle on
inlay hints if the server provides them.