Enable previously disabled arrow-keys
I disabled the arrow-keys to force myself to learn HJKL for movements.
However, I have a new keyboard (Keyboardio Model 100) that I'm going to
try to learn Colemak DH which moves HJKL.  It has a navigation layer
with arrow keys on the HJKL (for a qwerty keyboard), so by re-enabling
the arrow keys I get the best of both worlds.